+++ Products/CMFPlone/controlpanel/bbb/editing.py
@@ -1,13 +1,9 @@
-from zope.component import adapts
-from Products.CMFDefault.formlib.schema import ProxyFieldProperty
-from zope.interface import implements
-from zope.site.hooks import getSite
-from zope.component import getUtility
-from plone.registry.interfaces import IRegistry
-from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.interfaces import IPloneSiteRoot
-
 from Products.CMFPlone.interfaces import IEditingSchema
+from Products.CMFPlone.interfaces import IPloneSiteRoot
+from plone.registry.interfaces import IRegistry
+from zope.component import adapts
+from zope.component import getUtility
+from zope.interface import implements
 
 
 class EditingControlPanelAdapter(object):