+++ Products/CMFPlone/controlpanel/bbb/site.py
@@ -1,13 +1,10 @@
 # -*- coding: utf-8 -*-
 from Products.CMFDefault.formlib.schema import ProxyFieldProperty
 from Products.CMFPlone.interfaces import IPloneSiteRoot
-
 from Products.CMFPlone.interfaces import ISiteSchema
-
+from plone.registry.interfaces import IRegistry
 from zope.component import adapts
 from zope.component import getUtility
-from plone.registry.interfaces import IRegistry
-
 from zope.interface import implements