+++ Products/CMFPlone/controlpanel/tests/test_controlpanel_dateandtime.py
@@ -2,7 +2,6 @@ from Products.CMFPlone.interfaces import IDateAndTimeSchema
 from Products.CMFPlone.testing import PRODUCTS_CMFPLONE_INTEGRATION_TESTING
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
-
 import unittest2 as unittest