+++ Products/CMFPlone/controlpanel/tests/test_controlpanel_doctest.py
@@ -1,15 +1,11 @@
 # -*- coding: utf-8 -*-
 """Functional Doctests for control panel.
 """
-import doctest
-
-import unittest2 as unittest
-import pprint
-
+from Products.CMFPlone.testing import PRODUCTS_CMFPLONE_FUNCTIONAL_TESTING
 from plone.testing import layered
-
-from Products.CMFPlone.testing import \
-    PRODUCTS_CMFPLONE_FUNCTIONAL_TESTING
+import doctest
+import pprint
+import unittest2 as unittest
 
 
 optionflags = (
@@ -24,11 +20,12 @@ normal_testfiles = [
 def test_suite():
     suite = unittest.TestSuite()
     suite.addTests([
-        layered(doctest.DocFileSuite(test,
-                                     optionflags=optionflags,
-                                     globs={'pprint': pprint.pprint,
-                                            }
-                                     ),
-                layer=PRODUCTS_CMFPLONE_FUNCTIONAL_TESTING)
-        for test in normal_testfiles])
+        layered(
+            doctest.DocFileSuite(
+                test,
+                optionflags=optionflags,
+                globs={'pprint': pprint.pprint}
+            ),
+            layer=PRODUCTS_CMFPLONE_FUNCTIONAL_TESTING
+        ) for test in normal_testfiles])
     return suite