+++ Products/CMFPlone/skins/plone_ecmascript/inline_validation.js
@@ -44,8 +44,18 @@ jQuery(function ($) {
             fname = $field.attr('data-fieldname'),
             value = $input.val();
 
+        // value is null for empty multiSelection select, turn it into a [] instead
+        // so it does not break at_validate_field
+        if ($input.attr('multiple') === 'multiple' && value === null) {
+            value = $([]).serialize();
+        }
+
+        // if value is an Array, it will be send as value[]=value1&value[]=value2 by $.post
+        // turn it into something that will be useable or value will be omitted from the request
+        params = $.param({uid: uid, fname: fname, value: value}, traditional = true);
+
         if ($field && uid && fname) {
-            $.post($('base').attr('href') + '/at_validate_field', {uid: uid, fname: fname, value: value}, function (data) {
+            $.post($('base').attr('href') + '/at_validate_field', params, function (data) {
                 render_error($field, data.errmsg);
             });
         }